Simplifies usage, adds error handling, pep8 style
Simplifies usage so no input is required, songs just download.
Some changes are made to adhere to pep8 style guide in comments
and formatting. Downgraded to Python 2.7 because I couldn't get
ProgressBar to install for Python 3 but it seems to be working
for 2.7. Adds some basic error handling, was necessary to get
through my favorites list. Also added configurable (in code)
directory name where files are downlaoded to.